Backend Working

+++ src/main/java/com/example/springbootproject/model/Perdoruesi.java
@@ -2,15 +2,15 @@ package com.example.springbootproject.model;
 
 
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+
+import java.util.UUID;
 
 @Entity
 public class Perdoruesi {
 
 
-    private Integer id;
+    private long id;
     private String username;
     private String email;
     private String password;
@@ -40,11 +40,11 @@ public class Perdoruesi {
     }
 
     @Id
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
